+++ src/pages/api/oauth/kakao.ts
@@ -26,6 +26,21 @@ const signinRequest = async (code: string) => {
   return result;
 };
 
+/**
+ * - 로그인시 로직
+ * 1. 프론트에서 카카오에 인가 코드 요청을 보낸다.
+ * 2. 받은 인가 코드로 프론트 서버에서 백 서버에 로그인 요청을 보낸다.
+ * 3. 받은 엑세스 토큰을 쿠키에 저장한다.
+ * 4. 홈페이지로 리다이렉션 시킨다.
+ *
+ * - 회원가입시 로직
+ * 1. 프론트에서 카카오에 인가 코드 요청을 보낸다.
+ * 2. 받은 인가 코드로 프론트 서버에서 백 서버에 로그인 요청을 보낸다.
+ * 3. 로그인 실패시 간편 회원가입 페이지로 리다이렉션 시킨다.
+ * 4. 닉네임과 함께 인가 코드 요청을 다시 보낸다.
+ * 5. 받은 엑세스 토큰을 쿠키에 저장한다.
+ * 6. 홈페이지로 리다이렉션 시킨다.
+ */
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   const { code, state } = req.query;
 
@@ -62,21 +77,3 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
 };
 
 export default handler;
-
-/**
- * - 로그인시 로직
- * 1. 프론트에서 카카오에 인가 코드 요청을 보낸다.
- * 2. 받은 인가 코드로 프론트 서버에서 백 서버에 로그인 요청을 보낸다.
- * 3. 받은 엑세스 토큰을 쿠키에 저장한다.
- * 4. 홈페이지로 리다이렉션 시킨다.
- */
-
-/**
- * - 회원가입시 로직
- * 1. 프론트에서 카카오에 인가 코드 요청을 보낸다.
- * 2. 받은 인가 코드로 프론트 서버에서 백 서버에 로그인 요청을 보낸다.
- * 3. 로그인 실패시 간편 회원가입 페이지로 리다이렉션 시킨다.
- * 4. 닉네임과 함께 인가 코드 요청을 다시 보낸다.
- * 5. 받은 엑세스 토큰을 쿠키에 저장한다.
- * 6. 홈페이지로 리다이렉션 시킨다.
- */
