+++ src/pages/compare/index.tsx
@@ -4,7 +4,7 @@ import useGetInfiniteProducts from '@/models/product/useGetInfiniteProducts';
 import useProduct from '@/models/product/useProduct';
 import React, { SetStateAction, useEffect, useState } from 'react';
 import { Product } from '@/shared/types/product/product';
-import onClickCompare from '@/models/product/onClickCompare';
+import onClickCompare from '@/shared/utils/onClickCompare';
 import CompareResult from '@/shared/components/CompareResult/CompareResult';
 import CompareTable from '@/shared/components/CompareTable/CompareTable';
 import MogazoaLayout from '@/shared/components/App/MogazoaLayout';
