+++ src/shared/utils/compareProducts.ts
