+++ src/shared/utils/onClickCompare.ts
